# protos/gen/docs/go_app/v1/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "go_app/v1/api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Api"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/healthy": {
      "get": {
        "summary": "Health check",
        "operationId": "Api_Healthy",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Empty"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Api"
        ]
      }
    },
    "/ready": {
      "get": {
        "summary": "Readiness check",
        "operationId": "Api_Ready",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Empty"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Api"
        ]
      }
    },
    "/v1/todos": {
      "delete": {
        "operationId": "Api_DeleteTodos",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteTodosResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "tags": [
          "Api"
        ]
      },
      "post": {
        "operationId": "Api_UpsertTodos",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Todos"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UpsertTodosRequest"
            }
          }
        ],
        "tags": [
          "Api"
        ]
      }
    },
    "/v1/todos/get": {
      "post": {
        "operationId": "Api_GetTodos",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Todos"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "tags": [
          "Api"
        ]
      }
    },
    "/v1/todos/list": {
      "get": {
        "operationId": "Api_ListTodos",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Todos"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "orderBy",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Api"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1DeleteTodosResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "v1Empty": {
      "type": "object",
      "title": "Do not delete these unless you know what you're doing"
    },
    "v1Todo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "@gotags: fake:\"skip\""
        },
        "todo_type": {
          "$ref": "#/definitions/v1TodoType",
          "title": "@gotags: fake:\"{number:1,1}\""
        },
        "todo_name": {
          "type": "string",
          "title": "@gotags: fake:\"{hipsterword}\""
        },
        "priority": {
          "type": "integer",
          "format": "int32",
          "title": "@gotags: fake:\"{number:1,3}\""
        },
        "completed": {
          "type": "boolean",
          "title": "@gotags: fake:\"skip\""
        }
      }
    },
    "v1TodoType": {
      "type": "string",
      "enum": [
        "TODO_GENERIC",
        "TODO_SPECIFIC"
      ],
      "default": "TODO_GENERIC"
    },
    "v1Todos": {
      "type": "object",
      "properties": {
        "todos": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Todo"
          }
        }
      }
    },
    "v1UpsertTodosRequest": {
      "type": "object",
      "properties": {
        "todos": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Todo"
          }
        },
        "fields": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    }
  }
}